Update to be compatible with Connect 3.x
This latest version of Connect has moved the connect.static middleware
out of core.

This patch adds a dependency on the now-seperate "serve-static"
middleware and replaces the use of connect.static with that middleware
instead. This allows an install using "npm install" to actually work on
the first try--otherwise, users have to patch this. Alternate solutiosn
would be to pin the major version number in package.json to be for
connect 2.x